4월 29일 repository 수정

+++ trip/src/main/java/com/trip/entity/Planner/PlaceEntity.java
@@ -17,7 +17,7 @@ import lombok.Setter;
 public class PlaceEntity {
     @Id
     @GeneratedValue(strategy=GenerationType.IDENTITY)
-	private int placeId;
+	private Long placeId;
 	private String placeName;
 	
 	@ManyToOne(fetch=FetchType.LAZY)
